backend/requirements.txt
# FastAPI & Server
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18
websockets==14.1

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.7.0

# HeyGen / LiveAvatar
httpx==0.28.1

# LiveKit
livekit-api==0.7.1

# LLM Providers
openai==1.58.1
anthropic==0.40.0
litellm==1.55.8

# RAG & Embeddings
langchain==0.3.13
langchain-openai==0.3.0
langchain-anthropic==0.3.3
langchain-community==0.3.13
qdrant-client==1.12.1
chromadb==0.5.23

# Document Processing
pypdf==5.1.0
python-docx==1.1.2
beautifulsoup4==4.12.3
crawl4ai==0.4.2

# Utilities
python-dotenv==1.0.1
pydantic==2.10.3
redis==5.2.1
tenacity==9.0.0
structlog==24.4.0
